extensions: inline S3DeprovisionPipeline
