+++ extensions/aws/s3/provision/src/main/java/org/eclipse/dataspaceconnector/provision/aws/s3/S3BucketProvisioner.java
@@ -14,6 +14,7 @@
 
 package org.eclipse.dataspaceconnector.provision.aws.s3;
 
+import net.jodah.failsafe.Failsafe;
 import net.jodah.failsafe.RetryPolicy;
 import org.eclipse.dataspaceconnector.provision.aws.provider.ClientProvider;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
@@ -22,8 +23,25 @@ import org.eclipse.dataspaceconnector.spi.transfer.provision.ProvisionResponse;
 import org.eclipse.dataspaceconnector.spi.transfer.provision.Provisioner;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ProvisionedResource;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ResourceDefinition;
+import software.amazon.awssdk.services.iam.IamAsyncClient;
+import software.amazon.awssdk.services.iam.model.DeleteRolePolicyRequest;
+import software.amazon.awssdk.services.iam.model.DeleteRolePolicyResponse;
+import software.amazon.awssdk.services.iam.model.DeleteRoleRequest;
+import software.amazon.awssdk.services.iam.model.DeleteRoleResponse;
+import software.amazon.awssdk.services.s3.S3AsyncClient;
+import software.amazon.awssdk.services.s3.model.Delete;
+import software.amazon.awssdk.services.s3.model.DeleteBucketRequest;
+import software.amazon.awssdk.services.s3.model.DeleteBucketResponse;
+import software.amazon.awssdk.services.s3.model.DeleteObjectsRequest;
+import software.amazon.awssdk.services.s3.model.DeleteObjectsResponse;
+import software.amazon.awssdk.services.s3.model.ListObjectsV2Request;
+import software.amazon.awssdk.services.s3.model.ListObjectsV2Response;
+import software.amazon.awssdk.services.s3.model.ObjectIdentifier;
 
 import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
+
+import static java.util.stream.Collectors.joining;
 
 /**
  * Asynchronously provisions S3 buckets.
@@ -72,13 +90,57 @@ public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefiniti
     }
 
     @Override
-    public CompletableFuture<DeprovisionResponse> deprovision(S3BucketProvisionedResource provisionedResource) {
-        S3DeprovisionPipeline pipeline = S3DeprovisionPipeline.Builder.newInstance().clientProvider(clientProvider)
-                .retryPolicy(retryPolicy)
-                .monitor(monitor)
-                .resource().build();
+    public CompletableFuture<DeprovisionResponse> deprovision(S3BucketProvisionedResource resource) {
+        var s3Client = clientProvider.clientFor(S3AsyncClient.class, resource.getRegion());
+        var iamClient = clientProvider.clientFor(IamAsyncClient.class, resource.getRegion());
 
-        return pipeline.deprovision(provisionedResource);
+        String bucketName = resource.getBucketName();
+
+        monitor.info("S3 Deprovisioning: list bucket contents");
+        String role = resource.getRole();
+        return s3Client.listObjectsV2(listBucket(bucketName))
+                .thenCompose(listObjectsResponse -> deleteObjects(s3Client, bucketName, listObjectsResponse))
+                .thenCompose(deleteObjectsResponse -> deleteBucket(s3Client, bucketName))
+                .thenCompose(listAttachedRolePoliciesResponse -> deleteRolePolicy(iamClient, role))
+                .thenCompose(deleteRolePolicyResponse -> deleteRole(iamClient, role))
+                .thenApply(response -> DeprovisionResponse.Builder.newInstance().ok().resource(resource).build());
+    }
+
+    private CompletableFuture<DeleteRoleResponse> deleteRole(IamAsyncClient iamClient, String role) {
+        return Failsafe.with(retryPolicy).getStageAsync(() -> {
+            monitor.info("S3 Deprovisioning: delete role");
+            return iamClient.deleteRole(DeleteRoleRequest.builder().roleName(role).build());
+        });
+    }
+
+    private CompletableFuture<DeleteRolePolicyResponse> deleteRolePolicy(IamAsyncClient iamClient, String role) {
+        return Failsafe.with(retryPolicy).getStageAsync(() -> {
+            monitor.info("S3 Deprovisioning: deleting inline policies for Role " + role);
+            return iamClient.deleteRolePolicy(DeleteRolePolicyRequest.builder().roleName(role).policyName(role).build());
+        });
+    }
+
+    private CompletableFuture<DeleteBucketResponse> deleteBucket(S3AsyncClient s3Client, String bucketName) {
+        return Failsafe.with(retryPolicy).getStageAsync(() -> {
+            monitor.info("S3 Deprovisioning: delete bucket");
+            return s3Client.deleteBucket(DeleteBucketRequest.builder().bucket(bucketName).build());
+        });
+    }
+
+    private CompletableFuture<DeleteObjectsResponse> deleteObjects(S3AsyncClient s3Client, String bucketName, ListObjectsV2Response listObjectsResponse) {
+        var identifiers = listObjectsResponse.contents().stream()
+                .map(s3object -> ObjectIdentifier.builder().key(s3object.key()).build())
+                .collect(Collectors.toList());
+
+        var deleteRequest = DeleteObjectsRequest.builder()
+                .bucket(bucketName).delete(Delete.builder().objects(identifiers).build())
+                .build();
+        monitor.info("S3 Deprovisioning: delete bucket contents: " + identifiers.stream().map(ObjectIdentifier::key).collect(joining(", ")));
+        return s3Client.deleteObjects(deleteRequest);
+    }
+
+    private ListObjectsV2Request listBucket(String bucketName) {
+        return ListObjectsV2Request.builder().bucket(bucketName).build();
     }
 }
